Implementado o padrão fachada

+++ src/main/java/br/edu/unifei/ecot12/marinelife/Boat.java
@@ -1,13 +1,13 @@
 package br.edu.unifei.ecot12.marinelife;
 
 public class Boat extends Vehicle{
-    private float power;
+    private int power;
 
-    public float getPower() {
+    public int getPower() {
         return power;
     }
 
-    public void setPower(float power) {
+    public void setPower(int power) {
         this.power = power;
     }
 }
