+++ src/main/java/br/edu/unifei/ecot12/marinelife/Submarine.java
@@ -1,12 +1,12 @@
 package br.edu.unifei.ecot12.marinelife;
 
 public class Submarine extends Vehicle{
-    private float maximumDepth;
-    public float getMaximumDepth() {
+    private int maximumDepth;
+    public int getMaximumDepth() {
         return maximumDepth;
     }
 
-    public void setMaximumDepth(float maximumDepth) {
+    public void setMaximumDepth(int maximumDepth) {
         this.maximumDepth = maximumDepth;
     }
 }
